Added tests for create and delete
